added bunch of stuff

+++ 2025-11-18/MI-2023-2024-2/Z4.c
@@ -14,7 +14,7 @@ int main()
     for (int i = max_shift; i >= 0; i--)
     {
         int window = (number >> i) & mask;
-        if (window == sample)
+        if ((unsigned int)window == sample)
         {
             count++;
         }
